Allow the "primary" function to be overwritten
This does not add determinism to which plugin overwrites
the earlier function, just that an addition will overwrite
the primary implementation of the plugin.

Determining which plugin function should take priority is a deeply
complicated matter that should be addressed with significant research
and discussion.

At this time I think it should be discouraged overall, as it
introduces a system that could be used to replace models of
inheritence, and this should be strongly discouraged.